Moved video extension to _ext
